Backend/utopia/requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.6.0
fonttools==4.37.4
google-auth==2.16.0
google-auth-oauthlib==1.0.0
gspread==5.7.2
h11==0.14.0
idna==3.4
ipykernel==6.16.1
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.1
notebook_shim==0.2.0
numpy==1.23.3
oauthlib==3.2.2
outcome==1.2.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
phonenumbers==8.12.56
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.1
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psutil==5.9.3
psycopg2==2.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pycparser==2.21
pygame==2.1.2
Pygments==2.13.0
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.4
pywin32==304
pywinpty==2.0.8
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.7.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.1
terminado==0.16.0
tinycss2==1.2.1
tornado==6.2
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
twilio==7.14.2
tzdata==2022.1
urllib3==1.26.12
virtualenv==20.13.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
whitenoise==6.2.0
widgetsnbextension==4.0.3
wsproto==1.2.0
